Move Selection API to outline-helpers package
