+++ packages/outline-example/package.json
@@ -7,7 +7,8 @@
     "react-dom": "^17.0.1",
     "react-scripts": "4.0.0",
     "outline": "0.0.5",
-    "outline-react": "0.0.5"
+    "outline-react": "0.0.5",
+    "outline-helpers": "0.0.5"
   },
   "scripts": {
     "start": "react-scripts start",
