+++ packages/outline-react/package.json
@@ -15,7 +15,8 @@
   "version": "0.0.5",
   "dependencies": {
     "outline": "0.0.5",
-    "outline-extensions": "0.0.5"
+    "outline-extensions": "0.0.5",
+    "outline-helpers": "0.0.5"
   },
   "peerDependencies": {
     "react": ">=17.0.1",
