+++ packages/outline/src/OutlineUtils.js
@@ -12,7 +12,7 @@ export const emptyFunction = () => {};
 let keyCounter = 0;
 
 // inviariant(condition, message) will refine types based on "condition", and
-// if "condition" is false will throw an error. This functio is special-cased
+// if "condition" is false will throw an error. This function is special-cased
 // in flow itself, so we can't name it anything else.
 export function invariant(cond: boolean, message: string) {
   if (!cond) {
